Start a tesseract-plumbing crate
Initially, it is much the same as tesseract-rs, except that it deals in
cstrings. This removes a lot of `unwrap` calls.

It will become more different, and tesseract-rs will be rebuilt on top
of it.

The idea will be that tesseract-plumbing mirrors the tesseract c/c++ API
fairly closely, mostly making changes for safety and to avoid program
exit.

I don't know if I'll also end up with a leptonica-plumbing.

https://github.com/antimatter15/tesseract-rs/issues/17