Make it impossible to not initialize Tesseract
To recognise text, you first neeed to call initialize because it is the
only way of getting a TesseractInitialized object.

This avoids a situation where tesseract quits your program because you
forgot to initialize it.

Possibly the extra options will help with
https://github.com/antimatter15/tesseract-rs/issues/6.
I do plan on supporting all the options of
https://fossies.org/dox/tesseract-4.1.1/classtesseract_1_1TessBaseAPI.html#a96899e8e5358d96752ab1cfc3bc09f3e
in time.

https://github.com/antimatter15/tesseract-rs/issues/16